backend/requirements.txt
flask
pymongo
python-dotenv
dnspython
os